refactor(core): split CLI-ready core
