+++ src/core/integrity.ts
