+++ src/core/parser/SRecParser.ts
