+++ src/core/parser/segments.ts
