+++ src/webview/hexViewer.ts
@@ -10,7 +10,21 @@ import { renderInspector, renderBits, renderSegments, renderLabels, updateInspec
 import { renderStructPins, onSelectionChangeForStruct, resetStructViewState } from './struct';
 import { initSearch, runSearch, clearSearch, nextMatch, prevMatch } from './searchEngine';
 import { initFlatBytes, buildMemRows, getByte }      from './data';
-import type { SerializedParseResult, SerializedRecord, SidebarTab } from './types';
+import type { SerializedParseResult, SerializedRecord } from '../core/types';
+import type { SidebarTab } from './types';
+import {
+    crc8,
+    crc16,
+    crc32,
+    formatAnalyzeCommand,
+    formatAsciiByte,
+    formatCopyCommand,
+    formatHexArrayByte,
+    hexByte,
+    hexValue,
+    isAnalyzeCommand,
+    isCopyCommand,
+} from '../core/byte-tools';
 import { MAX_VIRTUAL_SCROLL_HEIGHT, physicalToLogicalScrollForLayout } from './virtualScroll';
 import {
     activateIntegrity,
@@ -1419,88 +1433,15 @@ function getSelBytes(): number[] {
     return out;
 }
 
-const COPY_COMMANDS = ['hex', 'hex-raw', 'binary', 'ascii', 'dec-array', 'hex-array', 'base64', 'dec', 'c-array'] as const;
-type CopyCommand = typeof COPY_COMMANDS[number];
-const COPY_COMMAND_SET = new Set<string>(COPY_COMMANDS);
-
-const COPY_FORMATTERS: Record<CopyCommand, (bytes: number[]) => string> = {
-    hex: bytes => bytes.map(hexByte).join(' '),
-    'hex-raw': bytes => bytes.map(hexByte).join(''),
-    binary: bytes => bytes.map(b => b.toString(2).padStart(8, '0')).join(' '),
-    ascii: bytes => bytes.map(formatAsciiByte).join(''),
-    'dec-array': bytes => `[${bytes.join(', ')}]`,
-    'hex-array': bytes => `[${bytes.map(formatHexArrayByte).join(', ')}]`,
-    base64: bytes => btoa(String.fromCharCode(...bytes)),
-    dec: bytes => `${bytes[0]}`,
-    'c-array': bytes => `{${bytes.map(formatHexArrayByte).join(', ')}}`,
-};
-
 function handleCopyCommand(cmd: string): void {
     const bytes = getSelBytes();
     if (bytes.length === 0 || !isCopyCommand(cmd)) { return; }
 
-    const text = COPY_FORMATTERS[cmd](bytes);
+    const text = formatCopyCommand(cmd, bytes);
     vscode.postMessage({ type: 'copyText', text, label: `${bytes.length} bytes as ${cmd}` });
 }
 
-function isCopyCommand(cmd: string): cmd is CopyCommand {
-    return COPY_COMMAND_SET.has(cmd);
-}
-
-function hexByte(b: number): string {
-    return b.toString(16).toUpperCase().padStart(2, '0');
-}
-
-function formatHexArrayByte(b: number): string {
-    return `0x${hexByte(b)}`;
-}
-
-function formatAsciiByte(b: number): string {
-    return (b >= 0x20 && b < 0x7F) ? String.fromCharCode(b) : '.';
-}
-
 // ── CRC helpers ──────────────────────────────────────────────────
-
-function crc8(data: number[]): number {
-    let c = 0;
-    for (const b of data) {
-        c ^= b;
-        for (let i = 0; i < 8; i++) { c = c & 0x80 ? ((c << 1) ^ 0x07) & 0xFF : (c << 1) & 0xFF; }
-    }
-    return c;
-}
-
-function crc16(data: number[]): number {
-    let c = 0;
-    for (const b of data) {
-        c ^= b;
-        for (let i = 0; i < 8; i++) { c = c & 1 ? ((c >>> 1) ^ 0xA001) : c >>> 1; }
-    }
-    return c & 0xFFFF;
-}
-
-function crc32(data: number[]): number {
-    let c = 0xFFFFFFFF;
-    for (const b of data) {
-        c ^= b;
-        for (let i = 0; i < 8; i++) { c = c & 1 ? ((c >>> 1) ^ 0xEDB88320) : c >>> 1; }
-    }
-    return (c ^ 0xFFFFFFFF) >>> 0;
-}
-
-type AnalyzeResult = { text: string; label: string };
-type AnalyzeFormatter = (bytes: number[]) => AnalyzeResult;
-const ANALYZE_COMMANDS = ['an-sum', 'an-xor', 'an-crc8', 'an-crc16', 'an-crc32'] as const;
-type AnalyzeCommand = typeof ANALYZE_COMMANDS[number];
-const ANALYZE_COMMAND_SET = new Set<string>(ANALYZE_COMMANDS);
-
-const ANALYZE_FORMATTERS: Record<AnalyzeCommand, AnalyzeFormatter> = {
-    'an-sum': formatAnalyzeSum,
-    'an-xor': formatAnalyzeXor,
-    'an-crc8': bytes => ({ text: `0x${hexValue(crc8(bytes))}`, label: 'CRC-8' }),
-    'an-crc16': bytes => ({ text: `0x${hexValue(crc16(bytes), 4)}`, label: 'CRC-16' }),
-    'an-crc32': bytes => ({ text: `0x${hexValue(crc32(bytes), 8)}`, label: 'CRC-32' }),
-};
 
 // ── Context menu ──────────────────────────────────────────────────
 
@@ -1524,24 +1465,9 @@ function handlePossibleFillCommand(cmd: string): void {
 function handleAnalyzeCommand(cmd: string, bytes: number[]): boolean {
     if (!isAnalyzeCommand(cmd)) { return false; }
 
-    const { text, label } = ANALYZE_FORMATTERS[cmd](bytes);
+    const { text, label } = formatAnalyzeCommand(cmd, bytes);
     vscode.postMessage({ type: 'copyText', text, label });
     return true;
-}
-
-function isAnalyzeCommand(cmd: string): cmd is AnalyzeCommand {
-    return ANALYZE_COMMAND_SET.has(cmd);
-}
-
-function formatAnalyzeSum(bytes: number[]): AnalyzeResult {
-    const sum = bytes.reduce((a, b) => a + b, 0);
-    const width = Math.max(4, sum.toString(16).length + (sum.toString(16).length % 2));
-    return { text: `0x${hexValue(sum, width)} (${sum})`, label: 'sum' };
-}
-
-function formatAnalyzeXor(bytes: number[]): AnalyzeResult {
-    const xor = bytes.reduce((a, b) => a ^ b, 0);
-    return { text: `0x${hexValue(xor)}`, label: 'XOR' };
 }
 
 function handleFillCommand(cmd: string): void {
@@ -1549,10 +1475,6 @@ function handleFillCommand(cmd: string): void {
 
     const val = parseInt(cmd.slice(5), 16);
     if (val >= 0 && val <= 0xFF) { applyFill(val); }
-}
-
-function hexValue(value: number, width = 2): string {
-    return value.toString(16).toUpperCase().padStart(width, '0');
 }
 
 function setupCtxMenu(): void {
